AdminOfficesIdmResourceTest test is extracted
